+++ src/main/kotlin/dev/furq/holodisplays/HoloDisplays.kt
@@ -26,7 +26,7 @@ import java.util.concurrent.Executors
 class HoloDisplays : ModInitializer {
     companion object {
         const val MOD_ID = "HoloDisplays"
-        const val VERSION = "0.4.2"
+        const val VERSION = "0.4.3"
         val LOGGER: Logger = LoggerFactory.getLogger(MOD_ID)
 
         var SERVER: MinecraftServer? = null